feat: Add new SVG icons to the project
Added new SVG icons to the project in the icons-sprites.svg file in the src/assets/images folder. Also, created a new Header component in the src/layout/header folder to display the header with navigation links. Updated the index.html file to link to the new favicon images and added a site.webmanifest file for metadata.